backend/requirements.txt
flask==2.3.3
qdrant-client==1.7.0
python-dotenv==1.0.0
flask-cors==4.0.0
openai==1.3.5
PyMuPDF==1.23.5
Pillow==10.0.1
langchain-openai==0.0.5
psycopg2-binary==2.9.7